+++ R/model_bootstrap.R
@@ -116,8 +116,7 @@
 #' mb_gam <- model_bootstrap(
 #'   attitude,
 #'   gam_attitude,
-#'   boot_it = 4,
-#'   parallel = 2  # CRAN limit (delete this line on your own computer)
+#'   boot_it = 4
 #' )
 #'
 #' # If the model is not standard, supply model_call_string with
@@ -130,8 +129,7 @@
 #'       raises + s(critical) + advance,
 #'     data = boot_data
 #'   )',
-#'   boot_it = 4,
-#'   parallel = 2  # CRAN limit (delete this line on your own computer)
+#'   boot_it = 4
 #' )
 #'
 #' # Model statistics and coefficients
@@ -139,8 +137,9 @@
 #' mb_gam$model_coefs
 #'
 #' # Plot ALE
-#' mb_gam$ale$plots |>
-#'   patchwork::wrap_plots()
+#' mb_gam_plots <- plot(mb_gam)
+#' mb_gam_1D_plots <- mb_gam_plots$distinct$rating$plots[[1]]
+#' patchwork::wrap_plots(mb_gam_1D_plots, ncol = 2)
 #' }
 #'
 model_bootstrap <- function (
@@ -490,8 +489,10 @@ model_bootstrap <- function (
               y_cat_actual
             }
             else {
+              # closeAllConnections()
+              # browser()
               # For numeric or ordinal data, actuals are the raw y_col values
-              data[oob_idxs, y_col] |>
+              data[oob_idxs, y_col, drop = FALSE] |>
                 pull()
             }